Move env.TOKEN_VALIDATION_FUNCTION extraction to index
All environment variables should be read in index, so it is easy to find
them all.

Also, move our default "isValid" definition to conf as the default value
for the environment variable. That way, we will always go through the same
code path making it more likely that we spot regressions in the "isValid"
setup code.

Also, add documentation to the auth code.